• add auto-indent support for public/private/protected storage modifiers, resulting in TM indenting classes as the following (and automatically decrease/increase indent when they are written):
    class my_class
    {
    public:
        my_class (arguments);
        virtual ~my_class ();

    private:
        /* data */
    };

I think this style is common?!? The only problem is that switch/case statements without braces around the case body will not indent properly. I *think* it did this in the past, but having to manually press tab after ‘case 32:’ if one decides not to use braces, seems like a little price to get the public/private/protected keywords to automatically indent when typed out.


git-svn-id: http://svn.textmate.org/trunk/Bundles/C.tmbundle@5311 dfb7d73b-c2ec-0310-8fea-fb051d288c6d