[NEW] Executor based “Build and Run” command for C/C++ that works like the Run command of Ruby/Python bundle.
Incidentally, the technique I used here could be useful for Java or other compiled languages.  We write our own bootstrap script to take the place of the interpreter.  The bootstrap script compiles the current file and runs the resulting executable.  This lets us use Executor unmodified, and we get all of executors features.



git-svn-id: http://svn.textmate.org/trunk/Bundles/C.tmbundle@10569 dfb7d73b-c2ec-0310-8fea-fb051d288c6d