Rewrite script to search for headers.
The script can now be used either with -p/--prefix to find a header with the given prefix, or with -s/--search to find a header with the given name.

There is also a -d/--domain parameter which can be set to ‘system’, ‘user’, or ‘all’, to select which domain to search.

Finally there is a new -C/--basedir parameter, this should be used when searching the user domain since the user domain includes the source’s directory (which this script doesn’t know about).

A change is that we now append the compiler supplied include paths to the user include paths (set via the TM_{USR,SYS}_HEADER_PATH environment variable).