Don’t increase indent for lines terminated with comma
This is lists like the following:

    NSUInteger modifiers[] = {
        NSShiftKeyMask,
        NSControlKeyMask,
        NSAlternateKeyMask,
        NSCommandKeyMask
    };

Despite the last item not being terminated by a comma, the above indents correctly as the last line decrease indent and “resets the carry”.

A minor issue with not indent on comma-terminated lines is function calls split over multiple lines. The previous behavior may have worked nicely for function calls split over only two lines, but beyond that, it didn’t work as desired, so I don’t think this is a big loss.